+++ ref-exchange/src/lib.rs
@@ -17,6 +17,7 @@ use crate::account_deposit::{VAccount, Account};
 pub use crate::action::SwapAction;
 use crate::action::{Action, ActionResult};
 use crate::errors::*;
+use crate::admin_fee::AdminFees;
 use crate::pool::Pool;
 use crate::simple_pool::SimplePool;
 use crate::utils::{check_token_duplicates};
@@ -25,6 +26,7 @@ pub use crate::views::{PoolInfo, ContractMetadata};
 mod account_deposit;
 mod action;
 mod errors;
+mod admin_fee;
 mod legacy;
 mod multi_fungible_token;
 mod owner;
@@ -109,9 +111,9 @@ impl Contract {
         self.internal_add_pool(Pool::SimplePool(SimplePool::new(
             self.pools.len() as u32,
             tokens,
-            fee + self.exchange_fee + self.referral_fee,
-            self.exchange_fee,
-            self.referral_fee,
+            fee,
+            0,
+            0,
         )))
     }
 
@@ -187,7 +189,11 @@ impl Contract {
         let mut amounts: Vec<u128> = amounts.into_iter().map(|amount| amount.into()).collect();
         let mut pool = self.pools.get(pool_id).expect("ERR_NO_POOL");
         // Add amounts given to liquidity first. It will return the balanced amounts.
-        pool.add_liquidity(&sender_id, &mut amounts);
+        pool.add_liquidity(
+            &sender_id,
+            &mut amounts,
+            AdminFees::new(self.exchange_fee),
+        );
         if let Some(min_amounts) = min_amounts {
             // Check that all amounts are above request min amounts in case of front running that changes the exchange rate.
             for (amount, min_amount) in amounts.iter().zip(min_amounts.iter()) {
@@ -220,6 +226,8 @@ impl Contract {
                 .into_iter()
                 .map(|amount| amount.into())
                 .collect(),
+            AdminFees::new(self.exchange_fee),
+            
         );
         self.pools.replace(pool_id, &pool);
         let tokens = pool.tokens();
@@ -335,8 +343,12 @@ impl Contract {
             amount_in,
             token_out,
             min_amount_out,
-            &self.owner_id,
-            referral_id,
+            AdminFees {
+                exchange_fee: self.exchange_fee,
+                exchange_id: env::current_account_id(),
+                referral_fee: self.referral_fee,
+                referral_id: referral_id.clone(),
+            },
         );
         self.pools.replace(pool_id, &pool);
         amount_out
@@ -358,7 +370,7 @@ mod tests {
     fn setup_contract() -> (VMContextBuilder, Contract) {
         let mut context = VMContextBuilder::new();
         testing_env!(context.predecessor_account_id(accounts(0)).build());
-        let contract = Contract::new(accounts(0), 4, 1);
+        let contract = Contract::new(accounts(0), 1600, 400);
         (context, contract)
     }
 
@@ -487,7 +499,7 @@ mod tests {
 
         // Get price from pool :0 1 -> 2 tokens.
         let expected_out = contract.get_return(0, accounts(1), one_near.into(), accounts(2));
-        assert_eq!(expected_out.0, 1662497915624478906119726);
+        assert_eq!(expected_out.0, 1663192997082117548978741);
 
         testing_env!(context
             .predecessor_account_id(accounts(3))
@@ -529,7 +541,7 @@ mod tests {
         // Exchange fees left in the pool as liquidity + 1m from transfer.
         assert_eq!(
             contract.get_pool_total_shares(0).0,
-            33337501041992301475 + 1_000_000
+            33336806279123620258 + 1_000_000
         );
 
         contract.withdraw(
@@ -773,8 +785,8 @@ mod tests {
             ],
             None,
         );
-        // Roundtrip returns almost everything except 0.3% fee.
-        assert_eq!(contract.get_deposit(acc, accounts(1)).0, 1_000_000 - 7);
+        // Roundtrip returns almost everything except 0.25% fee.
+        assert_eq!(contract.get_deposit(acc, accounts(1)).0, 1_000_000 - 6);
     }
 
     #[test]